making media and perf tabs toggleable
added two config options in the dashboard section to enable/disable the
media tab and the performance tab. if both are disabled, the tab bar in
the dashboard is hidden.

still a bit of jerkiness when config values resolve, not sure how to fix
that. also, the way i'm conditionally showing panels is kinda janky, i'm
sure there are better ways to do this...